+++ example/android/app/src/main/kotlin/com/shahxad/flutter_tex_example/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.shahxad.flutter_tex_example.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}