+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_tex/flutter_tex.dart';
 import 'package:flutter_tex_example/tex_view_document_example.dart';
@@ -7,7 +8,14 @@ import 'package:flutter_tex_example/tex_view_ink_well_example.dart';
 import 'package:flutter_tex_example/tex_view_markdown_example.dart';
 import 'package:flutter_tex_example/tex_view_quiz_example.dart';
 
-main() {
+main() async {
+  TeXRederingServer.renderingEngine = const TeXViewRenderingEngine.katex();
+
+  if (!kIsWeb) {
+    await TeXRederingServer.run();
+    await TeXRederingServer.initController();
+  }
+
   runApp(const FlutterTeXExample());
 }
 
@@ -33,14 +41,8 @@ class TeXViewFullExample extends StatefulWidget {
 class _TeXViewFullExampleState extends State<TeXViewFullExample> {
   int radVal = 0;
 
-  late TeXViewRenderingEngine renderingEngine;
-
   @override
   Widget build(BuildContext context) {
-    renderingEngine = radVal == 0
-        ? const TeXViewRenderingEngine.katex()
-        : const TeXViewRenderingEngine.mathjax();
-
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: AppBar(
@@ -57,68 +59,60 @@ class _TeXViewFullExampleState extends State<TeXViewFullExample> {
               height: 200,
             ),
           ),
-          RadioListTile<int>(
-            value: 0,
-            groupValue: radVal,
-            onChanged: (val) {
-              setState(() {
-                radVal = val!;
-              });
-            },
-            title: const Text("Katex"),
-            subtitle: const Text("RenderingEngine for Fast Rendering"),
-          ),
-          RadioListTile<int>(
-            value: 1,
-            groupValue: radVal,
-            onChanged: (val) {
-              setState(() {
-                radVal = val!;
-              });
-            },
-            title: const Text("MathJax"),
-            subtitle: const Text("RenderingEngine for Quality Rendering"),
+          Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              RadioListTile<int>(
+                value: 0,
+                groupValue: radVal,
+                onChanged: (val) async {
+                  TeXRederingServer.renderingEngine =
+                      const TeXViewRenderingEngine.katex();
+
+                  if (!kIsWeb) {
+                    await TeXRederingServer.initController();
+                  }
+
+                  setState(() {
+                    radVal = val!;
+                  });
+                },
+                title: const Text("Katex"),
+                subtitle: const Text("RenderingEngine for Fast Rendering"),
+              ),
+              RadioListTile<int>(
+                value: 1,
+                groupValue: radVal,
+                onChanged: (val) async {
+                  TeXRederingServer.renderingEngine =
+                      const TeXViewRenderingEngine.mathjax();
+                  if (!kIsWeb) {
+                    await TeXRederingServer.initController();
+                  }
+                  setState(() {
+                    radVal = val!;
+                  });
+                },
+                title: const Text("MathJax"),
+                subtitle: const Text("RenderingEngine for Quality Rendering"),
+              ),
+            ],
           ),
           const Divider(
             height: 30,
             color: Colors.transparent,
           ),
+          getExampleButton(context, 'Quiz Example', const TeXViewQuizExample()),
           getExampleButton(
-              context,
-              'Quiz Example',
-              TeXViewQuizExample(
-                renderingEngine: renderingEngine,
-              )),
+              context, 'TeX Examples', const TeXViewDocumentExamples()),
           getExampleButton(
-              context,
-              'TeX Examples',
-              TeXViewDocumentExamples(
-                renderingEngine: renderingEngine,
-              )),
+              context, 'Markdown Examples', const TeXViewMarkdownExamples()),
           getExampleButton(
-              context,
-              'Markdown Examples',
-              TeXViewMarkdownExamples(
-                renderingEngine: renderingEngine,
-              )),
+              context, 'Custom Fonts Examples', const TeXViewFontsExamples()),
+          getExampleButton(context, 'Image & Video Example',
+              const TeXViewImageVideoExample()),
           getExampleButton(
-              context,
-              'Custom Fonts Examples',
-              TeXViewFontsExamples(
-                renderingEngine: renderingEngine,
-              )),
-          getExampleButton(
-              context,
-              'Image & Video Example',
-              TeXViewImageVideoExample(
-                renderingEngine: renderingEngine,
-              )),
-          getExampleButton(
-              context,
-              'Inkwell Example',
-              TeXViewInkWellExample(
-                renderingEngine: renderingEngine,
-              )),
+              context, 'Inkwell Example', const TeXViewInkWellExample()),
         ],
       ),
     );