+++ example/lib/tex_view_fonts_example.dart
@@ -2,10 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_tex/flutter_tex.dart';
 
 class TeXViewFontsExamples extends StatelessWidget {
-  final TeXViewRenderingEngine renderingEngine;
 
   const TeXViewFontsExamples(
-      {super.key, this.renderingEngine = const TeXViewRenderingEngine.katex()});
+      {super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -16,22 +15,18 @@ class TeXViewFontsExamples extends StatelessWidget {
       ),
       body: TeXView(
           fonts: const [
-            TeXViewFont(fontFamily: 'army', src: 'assets/fonts/Army.ttf'),
-            TeXViewFont(fontFamily: 'budhrg', src: 'assets/fonts/Budhrg.ttf'),
-            TeXViewFont(fontFamily: 'celtg', src: 'assets/fonts/CELTG.ttf'),
-            TeXViewFont(fontFamily: 'hillock', src: 'assets/fonts/hillock.ttf'),
+            TeXViewFont(fontFamily: 'army', src: 'fonts/Army.ttf'),
+            TeXViewFont(fontFamily: 'budhrg', src: 'fonts/Budhrg.ttf'),
+            TeXViewFont(fontFamily: 'celtg', src: 'fonts/CELTG.ttf'),
+            TeXViewFont(fontFamily: 'hillock', src: 'fonts/hillock.ttf'),
+            TeXViewFont(fontFamily: 'intimacy', src: 'fonts/intimacy.ttf'),
             TeXViewFont(
-                fontFamily: 'intimacy', src: 'assets/fonts/intimacy.ttf'),
-            TeXViewFont(
-                fontFamily: 'sansation_light',
-                src: 'assets/fonts/SansationLight.ttf'),
-            TeXViewFont(
-                fontFamily: 'slenmini', src: 'assets/fonts/slenmini.ttf'),
+                fontFamily: 'sansation_light', src: 'fonts/SansationLight.ttf'),
+            TeXViewFont(fontFamily: 'slenmini', src: 'fonts/slenmini.ttf'),
             TeXViewFont(
                 fontFamily: 'subaccuz_regular',
-                src: 'assets/fonts/SubaccuzRegular.ttf'),
+                src: 'fonts/SubaccuzRegular.ttf'),
           ],
-          renderingEngine: renderingEngine,
           child: TeXViewColumn(children: [
             _teXViewWidget("Army", 'army'),
             _teXViewWidget("Budhrg", 'budhrg'),