+++ lib/src/utils/core_utils.dart
@@ -4,7 +4,7 @@ import 'package:flutter_tex/src/models/widget_meta.dart';
 import 'package:flutter_tex/src/utils/style_utils.dart';
 import 'package:flutter_tex/src/views/tex_view.dart';
 
-const double minHeight = 1;
+const double defaultHeight = 0.1;
 
 String getRawData(TeXView teXView) {
   return jsonEncode({